+++ Reisefradrag/services/IReisefradragBeregner.cs
@@ -4,6 +4,6 @@ namespace Reisefradrag.services
 {
     public interface IReisefradragBeregner
     {
-        public decimal BeregnFradrag(List<Arbeidsreise> arbeidsReiser, List<Besoeksreise> besoeksreiser, int utgifterBomFergeEtc);
+        public decimal BeregnFradrag(List<Arbeidsreise> arbeidsReiser, List<Besoeksreise> besoeksreiser, decimal utgifterBomFergeEtc);
     }
 }
